+++ src/main/java/qastudio/backend/global/security/oauth/OAuth2SuccessHandler.java
@@ -1,4 +1,4 @@
-package qastudio.backend.global.oauth;
+package qastudio.backend.global.security.oauth;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.Cookie;
